\name{shin}
          \alias{shin}
          \title{p53 Expression.}
          \description{
            This data set gives results about survival rates in patients with 
head and neck squamous cell carcinoma and the use of p-53 protein to predict 
them. 
          }
          \usage{data(shin)}
          \format{A table containing 16 columns and 20 rows.}
          \source{Data were extracted from Shin et al. 1996.}
          \references{Shin, D.M., Lee, J.S., Lippman, S.M., Lee, J.J., Tu, 
Z.N., Choi, G., Heyne, K., Shin, H.J.C., Ro, J.Y. Goepfert, H., Hong, W.K., and
Hittelman, W.N. (1996) "p53 Expression: Predicting Recurrence and Second 
Primary Tumors in Head and Neck Squamous Cell Carcinoma", Journal of National 
Cancer
Institute, 88, 519-529.
          }
          \keyword{datasets}



